move parameters enums as separate headers
- rng_type_t, sample_method_t, schedule_t, sd_type_t can be included without including the whole stable-diffusion.h file
- SDMode and SDParams are set in separate headers to highlight the main parameters of the API
