+++ examples/cli/CMakeLists.txt
@@ -1,6 +1,6 @@
 set(TARGET sd)
 
-add_executable(${TARGET} main.cpp)
+add_executable(${TARGET} main.cpp SDMode.hpp SDParams.hpp)
 install(TARGETS ${TARGET} RUNTIME)
 target_link_libraries(${TARGET} PRIVATE stable-diffusion ${CMAKE_THREAD_LIBS_INIT})
 target_compile_features(${TARGET} PUBLIC cxx_std_11)
