+++ examples/cli/SDMode.hpp
@@ -1,0 +1,12 @@
+#ifndef __SDMODE_HPP__
+#define __SDMODE_HPP__
+
+enum SDMode {
+    TXT2IMG,
+    IMG2IMG,
+    IMG2VID,
+    CONVERT,
+    MODE_COUNT
+};
+
+#endif  // __SDMODE_HPP__
